+++ software/reachy/io/io.py
@@ -22,3 +22,16 @@ class IO(object):
     def close(self):
         """Close and clean the IO."""
         raise NotImplementedError
+
+    def find_dual_camera(self, default_camera):
+        """Return a dual camera.
+
+        The camera object must implement following methods:
+            * read() -> bool, img
+            * close()
+            * set_active(side)
+
+        and the property:
+            * active_side --> 'left' or 'right'
+        """
+        raise NotImplementedError