+++ software/reachy/io/ws.py
@@ -6,6 +6,9 @@ import websockets
 import numpy as np
 
 from threading import Thread, Event
+from base64 import b64decode
+from io import BytesIO
+from PIL import Image
 
 from .io import IO
 
@@ -38,7 +41,14 @@ class WsIO(IO):
         For the moment no module are really implemented. Only placeholders for code compatibility are provided.
         """
         if module_name == 'force_gripper':
-            return WsFakeForceSensor()
+            force_sensor = WsFakeForceSensor()
+
+            if self.part_name == 'left_arm.hand':
+                self.ws.left_force_sensor = force_sensor
+            elif self.part_name == 'right_arm.hand':
+                self.ws.right_force_sensor = force_sensor
+
+            return force_sensor
 
         raise NotImplementedError
 
@@ -50,6 +60,7 @@ class WsIO(IO):
         pos = dxl_config['offset'] * (-1 if dxl_config['orientation'] == 'indirect' else 1)
         m = WsMotor(name=f'{self.part_name}.{dxl_name}', initial_position=pos)
         self.motors.append(m)
+        self.ws.motors[m.name] = m
         return m
 
     def find_orbita_disks(self):
@@ -61,6 +72,12 @@ class WsIO(IO):
         middleOrb = WsFakeOrbitaDisk()
         topOrb = WsFakeOrbitaDisk()
         return [bottomOrb, middleOrb, topOrb]
+
+    def find_dual_camera(self, default_camera):
+        """Retrieve a dual camera."""
+        cam = WsDualCamera(default_camera)
+        self.ws.cam = cam
+        return cam
 
     def close(self):
         """Close the WS."""
@@ -79,11 +96,7 @@ class WsMotor(object):
 
         self.compliant = False
         self.target_rot_position = initial_position
-
-    @property
-    def rot_position(self):
-        """Get the present position."""
-        return self.target_rot_position
+        self.rot_position = initial_position
 
 
 class WsFakeOrbitaDisk(object):
@@ -119,6 +132,32 @@ class WsFakeForceSensor(object):
         self.load = np.nan
 
 
+class WsDualCamera(object):
+    """Remote Camera."""
+
+    def __init__(self, default_camera):
+        """Set remote camera up."""
+        self.set_active(default_camera)
+        self.frame = np.zeros((300, 480, 3), dtype=np.uint8)
+
+    @property
+    def active_side(self):
+        """Get the active camera side."""
+        return self._camera_side
+
+    def set_active(self, camera_side):
+        """Set one of the camera active (left or right)."""
+        self._camera_side = camera_side
+
+    def read(self):
+        """Get latest received frame."""
+        return True, self.frame
+
+    def close(self):
+        """Close the camera."""
+        pass
+
+
 class WsServer(object):
     """WebSocket server, sync value from the modules with their equivalent from the client."""
 
@@ -128,6 +167,7 @@ class WsServer(object):
         self.running = Event()
 
         self.parts = []
+        self.motors = {}
 
     async def sync(self, websocket, path):
         """Sync loop that exchange modules state with the client."""
@@ -144,7 +184,24 @@ class WsServer(object):
                 ]
             })
             await websocket.send(msg.encode('UTF-8'))
-            await asyncio.sleep(0.01)
+
+            resp = await websocket.recv()
+            state = json.loads(resp)
+
+            if hasattr(self, 'cam'):
+                eye = f'{self.cam.active_side}_eye'
+                if eye in state:
+                    jpeg_data = b64decode(state[eye])
+                    self.cam.frame = np.array(Image.open(BytesIO(jpeg_data)))
+
+            for m in state['motors']:
+                if m['name'] in self.motors:
+                    self.motors[m['name']].rot_position = m['present_position']
+
+            if hasattr(self, 'left_force_sensor') and 'left_force_sensor' in state:
+                self.left_force_sensor.load = state['left_force_sensor']
+            if hasattr(self, 'right_force_sensor') and 'right_force_sensor' in state:
+                self.right_force_sensor.load = state['right_force_sensor']
 
     def close(self):
         """Stop the sync loop."""