+++ software/reachy/utils/__init__.py
@@ -7,4 +7,4 @@ from scipy.spatial.transform import Rotation as R
 
 def rot(axis, deg):
     """Compute 3D rotation matrix given euler rotation."""
-    return R.from_euler(axis, np.deg2rad(deg)).as_dcm()
+    return R.from_euler(axis, np.deg2rad(deg)).as_matrix()